+++ src/components/timelines.js
@@ -12,6 +12,7 @@ export default function Timelines() {
       duration: 1,
       repeat: 1,
       yoyo: true,
+      ease: "back",
     });
     tl.to(".purple", {
       x: 200,
@@ -20,6 +21,7 @@ export default function Timelines() {
       duration: 1,
       repeat: 1,
       yoyo: true,
+      ease: "bounce",
     });
     tl.to(".orange", {
       y: -200,
@@ -28,6 +30,7 @@ export default function Timelines() {
       rotation: 180,
       repeat: 1,
       yoyo: true,
+      ease: "bounce",
     });
     tl.to(".red", {
       x: 200,
@@ -36,16 +39,18 @@ export default function Timelines() {
       rotation: 180,
       repeat: 1,
       yoyo: true,
+      ease: "bounce",
     });
   }, []);
 
   return (
     <div className="main-div">
-      <div className="green"></div>
-      <div className="purple"></div>
-      <div className="orange"></div>
-      <div className="red"></div>
-      <div></div>
+      <div className="wrapper">
+        <div className="green"></div>
+        <div className="purple"></div>
+        <div className="orange"></div>
+        <div className="red"></div>
+      </div>
     </div>
   );
 }
